Refactor demo folder
