+++ extensions/sovity-messenger-demo/src/main/java/de/sovity/edc/extension/sovitymessenger/demo/SovityMessengerDemo.java
@@ -1,7 +1,5 @@
-package de.sovity.edc.extension.custommessage.receiver;
+package de.sovity.edc.extension.sovitymessenger.demo;
 
-import de.sovity.edc.extension.custommessage.receiver.message.Answer;
-import de.sovity.edc.extension.custommessage.receiver.message.Sqrt;
 import de.sovity.edc.extension.custommessages.api.MessageHandlerRegistry;
 import de.sovity.edc.extension.custommessages.api.PostOffice;
 import lombok.val;
