+++ extensions/sovity-messenger-demo/src/main/java/de/sovity/edc/extension/sovitymessenger/demo/Sqrt.java
@@ -1,4 +1,5 @@
-package de.sovity.edc.extension.custommessage.receiver.message;
+package de.sovity.edc.extension.sovitymessenger.demo;
+
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.sovity.edc.extension.custommessages.api.SovityMessage;
@@ -6,15 +7,16 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @Getter
-public class Answer implements SovityMessage {
+public class Sqrt implements SovityMessage {
+
     @Override
     public String getType() {
-        return "answer";
+        return "sqrt";
     }
 
-    @JsonProperty
-    private double answer;
+    @JsonProperty("value")
+    private double a;
 }
