+++ extensions/sovity-messenger-demo/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-de.sovity.edc.extension.custommessage.receiver.SovityMessengerDemo
+de.sovity.edc.extension.sovitymessenger.demo.SovityMessengerDemo
